changes 504 and migrant status to boolean. fixes administration condition codes in request.
