+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/DefaultReportOptionService.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.olap.service.impl;
 
 import org.opentestsystem.rdw.olap.model.CodedEntity;
 import org.opentestsystem.rdw.olap.model.ReportOptions;
+import org.opentestsystem.rdw.olap.repository.BooleanRepository;
 import org.opentestsystem.rdw.olap.repository.StrictBooleanRepository;
 import org.opentestsystem.rdw.olap.service.AdministrationConditionService;
 import org.opentestsystem.rdw.olap.service.AssessmentGradeService;
@@ -24,6 +25,7 @@ class DefaultReportOptionService implements ReportOptionService {
 
     private final AdministrationConditionService administrationConditionService;
     private final AssessmentGradeService assessmentGradeService;
+    private final BooleanRepository booleanRepository;
     private final CompletenessService completenessService;
     private final EthnicityService ethnicityService;
     private final GenderService genderService;
@@ -34,6 +36,7 @@ class DefaultReportOptionService implements ReportOptionService {
     DefaultReportOptionService(
             final AdministrationConditionService administrationConditionService,
             final AssessmentGradeService assessmentGradeService,
+            final BooleanRepository booleanRepository,
             final CompletenessService completenessService,
             final EthnicityService ethnicityService,
             final GenderService genderService,
@@ -42,8 +45,9 @@ class DefaultReportOptionService implements ReportOptionService {
             final SubjectService subjectService) {
         this.administrationConditionService = administrationConditionService;
         this.assessmentGradeService = assessmentGradeService;
-        this.ethnicityService = ethnicityService;
+        this.booleanRepository = booleanRepository;
         this.completenessService = completenessService;
+        this.ethnicityService = ethnicityService;
         this.genderService = genderService;
         this.schoolYearService = schoolYearService;
         this.strictBooleanRepository = strictBooleanRepository;
@@ -52,6 +56,7 @@ class DefaultReportOptionService implements ReportOptionService {
 
     @Override
     public ReportOptions getReportOptions() {
+        final List<CodedEntity> booleans = booleanRepository.findAll();
         final List<CodedEntity> strictBooleans = strictBooleanRepository.findAll();
         return ReportOptions.builder()
                 .assessmentGrades(assessmentGradeService.getAssessmentGrades())
@@ -63,9 +68,9 @@ class DefaultReportOptionService implements ReportOptionService {
                 .individualEducationPlans(strictBooleans)
                 .interimAdministrationConditions(administrationConditionService.getInterimConditions())
                 .limitedEnglishProficiencies(strictBooleans)
-                .migrantStatuses(strictBooleans)
+                .migrantStatuses(booleans)
                 .schoolYears(schoolYearService.getSchoolYears())
-                .section504s(strictBooleans)
+                .section504s(booleans)
                 .subjects(subjectService.getSubjects())
                 .summativeAdministrationConditions(administrationConditionService.getSummativeConditions())
                 .build();
