+++ webapp/src/main/webapp/src/app/aggregate-report/aggregate-report-request.mapper.ts
@@ -36,7 +36,7 @@ export class AggregateReportRequestMapper {
 
     if (!equalSize(settings.interimAdministrationConditions, options.interimAdministrationConditions)
       || !equalSize(settings.summativeAdministrationConditions, options.summativeAdministrationConditions)) {
-      query.administrationConditionCodes = settings.interimAdministrationConditions
+      query.administrativeConditionCodes = settings.interimAdministrationConditions
         .concat(settings.summativeAdministrationConditions);
     }
     if (!equalSize(settings.completenesses, options.completenesses)) {
